ai-matching/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
scikit-learn
numpy
